Refactor message factories for tests
