+++ tests/broadcast/repository_test.py
@@ -2,48 +2,18 @@
 
 from __future__ import annotations
 
-import arrow
 import pytest
 
-from semaphore.broadcast.models import (
-    BroadcastMessage,
-    FixedExpirationScheduler,
-    OpenEndedScheduler,
-    PermaScheduler,
-)
 from semaphore.broadcast.repository import (
     BroadcastMessageRepository,
     NotFoundError,
 )
 
-
-def create_active_message(identifier: str) -> BroadcastMessage:
-    return BroadcastMessage(
-        identifier=identifier,
-        summary_md="I'm an active message.",
-        body_md=None,
-        scheduler=PermaScheduler(),
-    )
-
-
-def create_future_message(identifier: str) -> BroadcastMessage:
-    start = arrow.utcnow().shift(hours=1)
-    return BroadcastMessage(
-        identifier=identifier,
-        summary_md="I'm an future message.",
-        body_md=None,
-        scheduler=OpenEndedScheduler(start),
-    )
-
-
-def create_stale_message(identifier: str) -> BroadcastMessage:
-    end = arrow.utcnow().shift(hours=-1)
-    return BroadcastMessage(
-        identifier=identifier,
-        summary_md="I'm an stale message.",
-        body_md=None,
-        scheduler=FixedExpirationScheduler(end),
-    )
+from ..support.broadcasts import (
+    create_active_message,
+    create_future_message,
+    create_stale_message,
+)
 
 
 def test_repository() -> None:
